add minimum version to readme
The minimum version is based on the version we explicitly test in CI, which is currently 1.21.0. It also outlines a policy for bumping the minimum version in minor releases if we need to. I think it's fair to say that we try hard not to bump that version unnecessarily, but I didn't want to get too defensive about it in the readme.